feat(avo-1912): split assignment response edit into 3 separate tabs
