+++ src/assignment/views/AssignmentResponseEdit/AssignmentResponseEdit.scss
